--- src/components/show-more.js
@@ -1,7 +1,0 @@
-/**
- * создаёт и возвращает разметку кнопки 'Показать больше'
- * @return {String}
- */
-export const getButtonShowMoreTemplate = () => (
-  `<button class="films-list__show-more">Show more</button>`
-);
